Cleans up some more tree code that is not used or required, moves the ui.treeEventHandler to a local var in the nav service since it doesn't need to be exposed, fixes main controller redirecting due to an undefined check.

+++ src/Umbraco.Web.UI.Client/src/views/common/main.controller.js
@@ -89,7 +89,7 @@ function MainController($scope, $rootScope, $location, $routeParams, $timeout, $
 
         //if the user has changed we need to redirect to the root so they don't try to continue editing the
         //last item in the URL
-        if (data.lastUserId !== undefined && data.lastUserId !== data.user.id) {
+        if (data.lastUserId !== undefined && data.lastUserId !== null && data.lastUserId !== data.user.id) {
             $location.path("/").search("");
             historyService.removeAll();
             treeService.clearCache();
